add menu despenser, nozzle & penerimaan

+++ database/migrations/2025_03_15_094615_create_reports_table.php
@@ -27,7 +27,7 @@ return new class extends Migration
             $table->integer('stok_aktual')->default(0);
             $table->integer('susut_pengeluaran')->default(0);
             $table->decimal('toleransi', 3, 1)->default(0.5);
-            $table->decimal('total_susut', 5, 2)->default(0);
+            $table->integer('total_susut')->default(0);
             $table->decimal('susut_harian', 5, 2)->default(0.0);
             $table->decimal('susut_bulanan', 5, 2)->default(0.0);
             $table->decimal('susut_tahunan', 5, 2)->default(0.0);
